requirements.txt
pandas
boto3
python-dotenv
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy